+++ rust/src/path.rs
@@ -23,7 +23,7 @@ impl<'a> Path<'a> {
     pub fn new<T: Contents>(repo: &'a Repo<T>, s: impl AsRef<str>) -> Result<Path, Error> {
         unsafe {
             let s = s.as_ref();
-            let ptr = irmin_path_of_string(repo.ptr, s.as_ptr() as *mut _, s.len() as i32);
+            let ptr = irmin_path_of_string(repo.ptr, s.as_ptr() as *mut _, s.len() as i64);
             if ptr.is_null() {
                 return Err(Error::NullPtr);
             }
@@ -72,7 +72,7 @@ impl<'a> Path<'a> {
                 self.repo.ptr,
                 self.ptr,
                 s.as_ptr() as *mut _,
-                s.len() as i32,
+                s.len() as i64,
             )
         };
         if ptr.is_null() {
@@ -87,12 +87,12 @@ impl<'a> Path<'a> {
 
     /// Convert a path to String
     pub fn to_string(&self) -> String {
-        let mut len = 0i32;
+        let mut len = 0u64;
         let ptr = unsafe { irmin_path_to_string(self.repo.ptr, self.ptr, &mut len) };
         if ptr.is_null() {
             return String::new();
         }
-        let s = IrminString(ptr, len);
+        let s = IrminString(ptr, len as usize);
         s.into()
     }
 }
